# src/main/webapp/docs/documentacion_wishes-1.0.0-swagger.yaml
openapi: 3.0.0
servers:
  - url: https://wishes-350515.oa.r.appspot.com/api
    description: API wishes, donde podemos encontar dos enlaces directos a books y bookslist
info:
  description: API to save all the things people like (books, films, music)
  version: 1.0.0
  title: Wishes
  contact:
    email: carruipor@alum.us.es
    name: Carmen Ruiz
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
  /books:
    get:
      tags:
        - books
      summary: Get books
      description: Get all books
      operationId: getBooks
      parameters:
        - name: order
          in: query
          description: Order results by genre, year, author, title, rate or page number
          required: false
          schema:
            type: string
            enum:
              - genre
              - year
              - author
              - title
              - rate
              - pagnumber
        - name: name
          in: query
          description: Query to filter. You can filter by title, author, genre, year, rate or publisher
          required: false
          schema:
            type: string
        - name: limit
          in: query
          description:  Maximum number of results
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: offset
          in: query
          description: Offset of the results
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        '200':
          description: Search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
        '400':
          description: Bad request
    post:
      tags:
        - books
      summary: Create a book
      description: Create a book
      operationId: createBook
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
        description: Song to create
        required: true
      responses:
        '201':
          description: Book created
          content:
            application/json:
              schema:
                 $ref: '#/components/schemas/Book'
        '400':
          description: Bad request
    put:
      tags:
        - books
      summary: Update a book
      description: Update a book
      operationId: updateBook
      requestBody:
        content:
          application/json:
            schema:
               $ref: '#/components/schemas/Book'
        description: Book to update
      responses:
        '204':
          description: Book updated
        '400':
          description: Bad request
        '404':
          description: Book not found
  '/books/{id}':
    get:
      tags:
        - books
      summary: Get a book
      description: Get a book
      operationId: getBook
      parameters:
        - name: id
          in: path
          description: Id of the book
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Book found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          description: Book not found
    delete:
      tags:
        - books
      summary: Delete book
      description: Delete book
      operationId: deleteBook
      parameters:
        - name: id
          in: path
          description: Id of the book
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Book deleted
        '404':
          description: Book not found
...
components:
  schemas:
    Book:
        type: object
        required:
          - title
        properties:
          id:
            type: string
          title:
            type: string
          author:
            type: string
          genre:
            type: string
          year:
            type: integer
          rate:
            type: number
          pagnumber:
            type: integer
          publisher:
            type: string
        example:
            id: b3
            title: Superfreakonomics
            author: Dubner, Stephen
            genre: economics
            year: 0
            rate: 0.
            pagnumber: 179
            publisher: HarperCollins
...